feat: create global isPlaying state and alpha tab hook
